framework/libraries/artie-tooling/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "artie-tooling"
readme = "README.md"
version = "0.0.1"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0"
]
